+++ lib/prompbmarshal/remote.pb.go
@@ -65,6 +65,7 @@ func sov(x uint64) (n int) {
 	return (bits.Len64(x|1) + 6) / 7
 }
 
+// IsEmpty checks if the WriteRequest has data to push.
 func (m *WriteRequest) IsEmpty() bool {
 	return m == nil || (len(m.Timeseries) == 0 && len(m.Metadata) == 0)
 }